+++ pkg/crch/R/CensoredStudentsT.R
@@ -65,3 +65,11 @@ support.CensoredStudentsT <- function(d, drop = TRUE, ...) {
   stopifnot(requireNamespace("distributions3"))
   distributions3::make_support(d$left, d$right, d, drop = drop)
 }
+
+is_discrete.CensoredStudentsT <- function(d, ...) {
+  setNames(rep.int(FALSE, length(d)), names(d))
+}
+
+is_continuous.CensoredStudentsT <- function(d, ...) {
+  setNames(!is.finite(d$left) & !is.finite(d$right), names(d))
+}